Crate debris_error

Compilation error handling

Exports the default Result type which is used throughout the compiler

Modules

lang_error 🔒

macros 🔒

This module contains some macros for the most common error messages

parse_error 🔒

snippet 🔒

Owned variants of the annotate_snippets library structs

utils 🔒

Macros

debris_unimplemented

Structs

CompileErrors

LangError

A generic error which gets thrown when compiling

SliceOwned

An owned counterpart to the annotate_snippets::Slice struct

SnippetOwned

An owned counterpart to the annotate_snippets::Snippet struct

SourceAnnotationOwned

An owned counterpart to the annotate_snippets::SourceAnnotation struct

Enums

ControlFlowRequires

LangErrorKind

Specifies a specific error reason

ParseError

Thrown when parsing bad input

SingleCompileError

A Compile Error

Constants

COLORED

Used to determine whether the error messages should use console colors Or be plain text

Traits

AsAnnotationSnippet

Type Definitions

LangResult

A result type which allows quick error throwing since no span and other boilerplate is needed

Result

The result type used by most of the core functions